travelsBackend/requirements.txt
backports.shutil-get-terminal-size==1.0.0
cairocffi==0.6
CairoSVG==1.0.22
cffi==1.2.1
cssselect==0.9.2
decorator==4.0.10
Django==1.7
django-crispy-forms==1.6.0
django-filter==0.7
django-guardian==1.4.2
django-model-utils==2.5
djangorestframework==3.3.3
djoser==0.4.3
drf-extensions==0.3.1
drf-tracking==0.2.5
drfdocs==0.0.9
enum34==1.1.6
html5lib==0.999999999
inflection==0.3.1
ipython==5.1.0
ipython-genutils==0.1.0
lxml==3.6.4
Markdown==2.6.6
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
pycparser==2.14
Pygments==2.1.3
Pyphen==0.9.4
python-memcached==1.58
pytz==2016.3
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
tinycss==0.3
traitlets==4.3.0
wcwidth==0.1.7
WeasyPrint==0.31
webencodings==0.5
workdays==1.4

